refactor: Refactored layout component
