+++ lib/api/index.ts
@@ -2,3 +2,4 @@ export * from './posts';
 export * from './newsletter';
 export * from './strapi';
 export * from './collection';
+export * from './layout';
