+++ lib/api/layout/index.ts
@@ -1,0 +1,69 @@
+import { assertNever, Awaited, filterNullableItems } from '@lib/utils';
+import { LayoutPropsFragment } from '../strapi/generated/sdk';
+import { getImageFromStrapiImage } from '../utils';
+
+type StoreSettings = NonNullable<
+   NonNullable<LayoutPropsFragment['storeSettings']>[0]
+>;
+
+type Footer = NonNullable<StoreSettings['footer']>;
+
+type Menu = NonNullable<Footer['menu1']>;
+
+export type LayoutData = NonNullable<
+   Awaited<ReturnType<typeof getLayoutProps>>
+>['layout'];
+
+export function getLayoutProps(data: LayoutPropsFragment) {
+   const storeSettings = data.storeSettings?.[0];
+   const footer = storeSettings?.footer;
+   const socialMediaAccounts = storeSettings?.socialMediaAccounts || {};
+   const stores = filterNullableItems(data.stores);
+   return {
+      layout: {
+         footer: {
+            menu1: footer?.menu1 ? getMenu(footer.menu1) : undefined,
+            menu2: footer?.menu2 ? getMenu(footer.menu2) : undefined,
+            partners: footer?.partners ? getMenu(footer.partners) : undefined,
+            bottomMenu: footer?.bottomMenu
+               ? getMenu(footer.bottomMenu)
+               : undefined,
+            socialMediaAccounts,
+            stores,
+         },
+      },
+   };
+}
+
+function getMenu(rawMenu: Menu) {
+   return {
+      items: filterNullableItems(rawMenu.items).map((item) => {
+         switch (item.__typename) {
+            case 'ComponentMenuLink': {
+               return {
+                  name: item.name,
+                  url: item.url,
+               };
+            }
+            case 'ComponentMenuLinkWithImage': {
+               return {
+                  name: item.name,
+                  url: item.url,
+                  image:
+                     getImageFromStrapiImage(item.image, 'small') || undefined,
+               };
+            }
+            case 'ComponentMenuCollectionLink': {
+               return {
+                  name: item.name,
+                  url: item.linkedCollection
+                     ? `/collections/${item.linkedCollection.handle}`
+                     : '#',
+               };
+            }
+            default:
+               return assertNever(item);
+         }
+      }),
+   };
+}
