+++ lib/api/strapi/operations/layout.graphql
@@ -1,0 +1,56 @@
+fragment LayoutProps on Query {
+   stores {
+      code
+      name
+      url
+      currency
+   }
+   storeSettings(where: $whereStoreSettings) {
+      footer {
+         menu1 {
+            ...MenuProps
+         }
+         menu2 {
+            ...MenuProps
+         }
+         partners {
+            ...MenuProps
+         }
+         bottomMenu {
+            ...MenuProps
+         }
+      }
+      socialMediaAccounts {
+         twitter
+         facebook
+         instagram
+         youtube
+         repairOrg
+      }
+   }
+}
+
+fragment MenuProps on Menu {
+   items {
+      __typename
+      ... on ComponentMenuLink {
+         name
+         url
+      }
+      ... on ComponentMenuCollectionLink {
+         name
+         linkedCollection {
+            handle
+         }
+      }
+      ... on ComponentMenuLinkWithImage {
+         name
+         url
+         image {
+            alternativeText
+            url
+            formats
+         }
+      }
+   }
+}
