+++ pages/_app.tsx
@@ -1,11 +1,21 @@
 import { AppProviders } from '@components/AppProviders';
+import { NextPage } from 'next';
 import { AppProps } from 'next/app';
 import * as React from 'react';
 
-function MyApp({ Component, pageProps }: AppProps) {
+type NextPageWithLayout = NextPage & {
+   getLayout?: (page: React.ReactElement, pageProps: any) => React.ReactNode;
+};
+
+type AppPropsWithLayout = AppProps & {
+   Component: NextPageWithLayout;
+};
+
+function MyApp({ Component, pageProps }: AppPropsWithLayout) {
+   const getLayout = Component.getLayout ?? ((page) => page);
    return (
       <AppProviders>
-         <Component {...pageProps} />
+         {getLayout(<Component {...pageProps} />, pageProps)}
       </AppProviders>
    );
 }
